Added linux example to imgui_app
* Updated sokol (for some reason it wasn't working on linux)

+++ src/imgui_app.cpp
@@ -93,6 +93,6 @@ ImTextureID imgui_app_loadImageRGBA8(const void *data, int width, int height) {
 }
 
 void imgui_app_destroyImage(ImTextureID id) {
-    sg_image img = {(uint32_t) id};
+    sg_image img = {(uint32_t)(size_t)id};
     sg_destroy_image(img);
 }
